@@ -4,10 +4,10 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "relecov-tools"
-version = "1.5.5"
+version = "1.6.0"
 description = "Tools for managing and proccessing relecov network data."
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 authors = [
     {name = "Sara Monzon", email = "smonzon@isciii.es"},
     {name = "Luis Chapado", email = "lchapado@externos.isciii.es"},
