@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 import os
-import sys
 import rich.console
 import re
 import shutil
+import importlib
 import numpy as np
 from datetime import datetime
 from rich.prompt import Prompt
@@ -77,6 +77,7 @@ class BioinfoMetadata(BaseModule):
         output_dir=None,
         software_name=None,
         update=False,
+        soft_validation=False,
         **kwargs,
     ):
         super().__init__(output_dir=output_dir, called_module=__name__)
@@ -112,9 +113,8 @@ class BioinfoMetadata(BaseModule):
                 "error",
                 f"file {json_file} does not exist",
             )
-            sys.exit(
-                self.log_report.print_log_report(self.__init__.__name__, ["error"])
-            )
+            self.log_report.print_log_report(self.__init__.__name__, ["error"])
+            raise ValueError(f"Json file {json_file} does not exist, cannot continue.")
 
         #  Assign a new name for better readability
         self.readlabmeta_json_file = json_file
@@ -124,14 +124,8 @@ class BioinfoMetadata(BaseModule):
         self.j_data = self.collect_info_from_lab_json()
         batch_id = self.get_batch_id_from_data(self.j_data)
         self.set_batch_id(batch_id)
-
-        meta_basename = os.path.basename(self.readlabmeta_json_file).split(".")[0]
-        if batch_id in meta_basename:
-            # File is probably tagged already from a previous process
-            self.out_filename = "bioinfo_" + meta_basename + ".json"
-        else:
-            self.out_filename = self.tag_filename("bioinfo_" + meta_basename + ".json")
         self.update = update
+        self.soft_validation = soft_validation
 
         # Parse input/output folder
         if input_folder is None:
@@ -163,14 +157,14 @@ class BioinfoMetadata(BaseModule):
         if self.software_name in available_software:
             self.software_config = bioinfo_config.get_configuration(self.software_name)
         else:
+            errtxt = f"No configuration available for '{self.software_name}'. Currently, the only available software options are:: {', '.join(available_software)}"
             self.update_all_logs(
                 self.__init__.__name__,
                 "error",
-                f"No configuration available for '{self.software_name}'. Currently, the only available software options are:: {', '.join(available_software)}",
+                errtxt,
             )
-            sys.exit(
-                self.log_report.print_log_report(self.__init__.__name__, ["error"])
-            )
+            self.log_report.print_log_report(self.__init__.__name__, ["error"])
+            raise ValueError(errtxt)
 
     def update_all_logs(self, method_name, status, message):
         report = self.log_report.update_log_report(method_name, status, message)
@@ -214,12 +208,14 @@ class BioinfoMetadata(BaseModule):
                         files_found[topic_key] = []
                     files_found[topic_key].extend(matching_files)
         if len(files_found) < 1:
+            errtxt = f"No files found in '{self.input_folder}' according to '{os.path.basename(self.bioinfo_json_file)}' file name patterns."
             self.update_all_logs(
                 method_name,
                 "error",
-                f"No files found in '{self.input_folder}' according to '{os.path.basename(self.bioinfo_json_file)}' file name patterns.",
+                errtxt,
             )
-            sys.exit(self.log_report.print_log_report(method_name, ["error"]))
+            self.log_report.print_log_report(method_name, ["error"])
+            raise ValueError(errtxt)
         else:
             self.update_all_logs(
                 self.scann_directory.__name__,
@@ -394,12 +390,14 @@ class BioinfoMetadata(BaseModule):
             else:
                 continue
         if len(missing_required) >= 1:
+            errtxt = f"Missing mandatory files in {self.software_name}:{', '.join(missing_required)}"
             self.update_all_logs(
                 method_name,
                 "error",
-                f"Missing mandatory files in {self.software_name}:{', '.join(missing_required)}",
+                errtxt,
             )
-            sys.exit(self.log_report.print_log_report(method_name, ["error"]))
+            self.log_report.print_log_report(method_name, ["error"])
+            raise ValueError(errtxt)
         else:
             self.update_all_logs(
                 method_name, "valid", "Successfull validation of mandatory files."
@@ -408,7 +406,7 @@ class BioinfoMetadata(BaseModule):
         return
 
     def add_bioinfo_results_metadata(
-        self, files_dict, j_data, sufix, file_tag, output_dir=None
+        self, files_dict, file_tag, j_data, output_dir=None
     ):
         """Adds metadata from bioinformatics results to j_data.
         It first calls file_handlers and then maps the handled
@@ -417,15 +415,18 @@ class BioinfoMetadata(BaseModule):
         Args:
             files_dict (dict{str:str}): A dictionary containing file paths found based on the definitions provided in the bioinformatic JSON file within the software scope (self.software_config).
             j_data (list(dict{str:str}): A list of dictionaries containing metadata lab (list item per sample).
-            sufix (str): Sufix added to splitted tables file name.
             output_dir (str): Path to save output files generated during handling_files() process.
-            file_tag(str): Tag that will be used for output filenames includes batch date (same as download date) and hex.
+            file_tag (str): Tag that will be used for output filenames includes batch date (same as download date) and hex.
 
         Returns:
             j_data_mapped: A list of dictionaries with bioinformatics metadata mapped into j_data.
+            extra_json_data: List of dictionaries with data that hasn't to be mapped in order to be processed afterwards.
         """
         method_name = f"{self.add_bioinfo_results_metadata.__name__}"
+        extra_json_data = []
         for key in self.software_config.keys():
+            # Reset map_data flag so it only activates when table expects mapping
+            map_data_flag = False
             # Update bioinfo cofiguration key/scope
             self.current_config_key = key
             map_method_name = f"{method_name}:{self.software_name}.{key}"
@@ -443,15 +444,27 @@ class BioinfoMetadata(BaseModule):
                     f"No file path found for '{self.software_name}.{key}'",
                 )
                 continue
-            # Handling files
-            if self.software_config[key].get("map", True) is False:
-                msg = f"File '{self.software_name}.{key}' was processed but skipped from mapping as defined in config."
-                self.update_all_logs(map_method_name, "warning", msg)
-                self.log_report.print_log_report(map_method_name, ["warning"])
-                continue
-            data_to_map = self.handling_files(
-                files_dict[key], sufix, output_dir, file_tag
-            )
+
+            current_config = self.software_config[self.current_config_key]
+            file_name = current_config.get("fn")
+            func_name = current_config.get("function")
+
+            if func_name is None:
+                data = self.handling_tables(files_dict[key], file_name)
+            else:
+                data = self.process_metadata(
+                    files_dict[key], file_tag, func_name, output_dir
+                )
+
+            if current_config.get("split_by_batch") and current_config.get(
+                "extra_dict"
+            ):
+                extra_json_data.append(data)
+                map_data_flag = False
+            else:
+                data_to_map = data
+                map_data_flag = True
+
             # Mapping data to j_data
             mapping_fields = self.software_config[key].get("content")
             if not mapping_fields:
@@ -462,7 +475,7 @@ class BioinfoMetadata(BaseModule):
                 )
                 self.log_report.print_log_report(map_method_name, ["warning"])
                 continue
-            if data_to_map:
+            if data_to_map and map_data_flag:
                 j_data_mapped = self.mapping_over_table(
                     j_data=j_data,
                     map_data=data_to_map,
@@ -477,7 +490,7 @@ class BioinfoMetadata(BaseModule):
                 )
                 continue
         self.log_report.print_log_report(method_name, ["valid", "warning"])
-        return j_data_mapped
+        return j_data_mapped, extra_json_data
 
     def handling_tables(self, file_list, conf_tab_name):
         """Reads a tabular file in different formats and returns a dictionary containing
@@ -533,125 +546,51 @@ class BioinfoMetadata(BaseModule):
             )
             raise ValueError(self.log_report.print_log_report(method_name, ["error"]))
 
-    def handling_files(self, file_list, sufix, output_dir, file_tag):
-        """Handles different file formats to extract data regardless of their structure.
-        The goal is to extract the data contained in files specified in ${file_list},
-        using either 'standard' handlers defined in this class or pipeline-specific file handlers.
-        (inspired from ./metadata_homogenizer.py)
-
-            A file handler method must generate a data structure as follow:
-            {
-                'SAMPLE1': {
-                    'field1': 'value1'
-                    'field2': 'value2'
-                    'field3': 'value3'
-                },
-                SAMPLE2': {
-                    'field1': 'value1'
-                    'field2': 'value2'
-                    'field3': 'value3'
-                },
-                ...
-            }
-            Note: ensure that 'field1','field2','field3' corresponds with the values
-            especified in the 'content' section of each software configuration scope
-            (see: conf/bioinfo_config.json).
+    def process_metadata(self, file_list, file_tag, func_name, out_path):
+        """This method dynamically loads and executes the functions especified in config file.
+        It is used to apply standard or custom metadata processing depending on the current
+        software context (`self.software_name`).
 
         Args:
             file_list (list): A list of file path/s to be processed.
-            output_dir (str): Path to save output files from imported method if necessary
-            file_tag(str): Tag that will be used for output filenames includes batch date (same as download date) and hex.
+            func_name (str): The name of the function to execute
+            file_tag (str): Tag that will be used for output filenames includes batch date (same as download date) and hex.
+            out_path (str): Path to save output files generated during handling_files() process.
 
         Returns:
             data: A dictionary containing bioinfo metadata handled for each sample.
         """
-        method_name = f"{self.add_bioinfo_results_metadata.__name__}:{self.handling_files.__name__}"
-        splitted_path = os.path.join(output_dir, "analysis_results")
-        file_name = self.software_config[self.current_config_key].get("fn")
-        # Parsing files
-        current_config = self.software_config[self.current_config_key]
-        func_name = current_config.get("function")
-        if func_name is None:
-            data = self.handling_tables(file_list=file_list, conf_tab_name=file_name)
-        else:
-            if current_config.get("split_by_batch") is True:
-                file_extension = current_config.get("fn").rsplit(".", 1)[1]
-                base_filename = current_config.get("fn").rsplit(".", 1)[0]
-                pattern = re.compile(
-                    f"{base_filename}_{sufix}.{re.escape(file_extension)}"
-                )
-                matching_files = [
-                    f for f in os.listdir(splitted_path) if pattern.match(f)
-                ]
-                full_paths = [  # noqa: F841
-                    os.path.join(splitted_path, f) for f in matching_files
-                ]
-                try:
-                    if func_name.startswith("utils/"):
-                        utils_name = "relecov_tools.assets.pipeline_utils.utils"
-                        func_name = func_name.split("/", 1)[1]
-                    else:
-                        utils_name = (
-                            f"relecov_tools.assets.pipeline_utils.{self.software_name}"
-                        )
-
-                    # Dynamically import the function from the specified module
-                    import_statement = f"import {utils_name}"
-                    exec(import_statement)
-                    # Get method name and execute it.
-                    data = eval(
-                        utils_name
-                        + "."
-                        + func_name
-                        + "(full_paths, file_tag, '"
-                        + self.software_name
-                        + "', output_dir)"
-                    )
-
-                except Exception as e:
-                    self.update_all_logs(
-                        self.save_merged_files.__name__,
-                        "error",
-                        f"Error occurred while parsing '{func_name}': {e}.",
-                    )
-                    sys.exit(self.log_report.print_log_report(method_name, ["error"]))
+        method_name = f"{self.add_bioinfo_results_metadata.__name__}:{self.process_metadata.__name__}"
+        try:
+            # Dynamically import the function from the specified module
+            if func_name.startswith("utils/"):
+                utils_name = "relecov_tools.assets.pipeline_utils.utils"
+                func_name = func_name.split("/", 1)[1]
             else:
-                try:
-                    # Dynamically import the function from the specified module
-                    if func_name.startswith("utils/"):
-                        utils_name = "relecov_tools.assets.pipeline_utils.utils"
-                        func_name = func_name.split("/", 1)[1]
-                    else:
-                        utils_name = (
-                            f"relecov_tools.assets.pipeline_utils.{self.software_name}"
-                        )
+                utils_name = f"relecov_tools.assets.pipeline_utils.{self.software_name}"
 
-                    # Dynamically import the function from the specified module
-                    import_statement = f"import {utils_name}"
-                    exec(import_statement)
-                    # Get method name and execute it.
-                    data = eval(
-                        utils_name
-                        + "."
-                        + func_name
-                        + "(file_list, file_tag, '"
-                        + self.software_name
-                        + "', output_dir)"
-                    )
-                except Exception as e:
-                    self.update_all_logs(
-                        self.add_bioinfo_results_metadata.__name__,
-                        "error",
-                        f"Error occurred while parsing '{func_name}': {e}.",
-                    )
-                    sys.exit(self.log_report.print_log_report(method_name, ["error"]))
+            # Dynamically import the function from the specified module
+            module = importlib.import_module(utils_name)
+            # Get method from func_name and execute it.
+            func_obj = getattr(module, func_name)
+            data = func_obj(file_list, file_tag, self.software_name, out_path)
+
+        except Exception as e:
+            self.update_all_logs(
+                method_name,
+                "error",
+                f"Error occurred while parsing '{func_name}': {e}.",
+            )
+            self.log_report.print_log_report(method_name, ["error"])
+            raise ValueError(f"Error occurred while parsing '{func_name}': {e}.")
         return data
 
-    def add_fixed_values(self, j_data):
+    def add_fixed_values(self, j_data, out_filename):
         """Add fixed values to j_data as defined in the bioinformatics configuration (definition: "fixed values")
 
         Args:
             j_data (list(dict{str:str}): A list of dictionaries containing metadata lab (one item per sample).
+            out_filename (str): File name of the bioinfo_lab_metadata json
 
         Returns:
             j_data: updated j_data with fixxed values added in it.
@@ -660,7 +599,7 @@ class BioinfoMetadata(BaseModule):
         try:
             f_values = self.software_config["fixed_values"]
             for row in j_data:
-                row["bioinfo_metadata_file"] = self.out_filename
+                row["bioinfo_metadata_file"] = out_filename
                 for field, value in f_values.items():
                     row[field] = value
             self.update_all_logs(method_name, "valid", "Fields added successfully.")
@@ -739,11 +678,21 @@ class BioinfoMetadata(BaseModule):
                             self.software_config[key].get("extract")
                             or self.software_config[key].get("function")
                         ):
-                            analysis_results_path = os.path.join(
-                                base_cod_path,
-                                "analysis_results",
-                                os.path.basename(paths),
-                            )
+                            if self.software_config[key].get("split_by_batch"):
+                                base, ext = os.path.splitext(os.path.basename(paths))
+                                batchid = row["batch_id"]
+                                new_fname = f"{base}_{batchid}_{self.hex}{ext}"
+                                analysis_results_path = os.path.join(
+                                    base_cod_path,
+                                    "analysis_results",
+                                    new_fname,
+                                )
+                            else:
+                                analysis_results_path = os.path.join(
+                                    base_cod_path,
+                                    "analysis_results",
+                                    os.path.basename(paths),
+                                )
                             analysis_results_paths.append(analysis_results_path)
                         else:
                             analysis_results_paths = file_path
@@ -777,12 +726,16 @@ class BioinfoMetadata(BaseModule):
                 self.readlabmeta_json_file
             )
         except ValueError:
-            self.update_all_logs(
-                self.collect_info_from_lab_json.__name__,
-                "error",
-                f"Invalid lab-metadata json file: self.{self.readlabmeta_json_file}",
+            errtxt = (
+                f"Invalid lab-metadata json file: self.{self.readlabmeta_json_file}"
             )
-            sys.exit(self.log_report.print_log_report(method_name, ["error"]))
+            self.update_all_logs(
+                method_name,
+                "error",
+                errtxt,
+            )
+            self.log_report.print_log_report(method_name, ["error"])
+            raise ValueError(errtxt)
         return json_lab_data
 
     def get_sample_idx_colpos(self, config_key):
@@ -840,7 +793,6 @@ class BioinfoMetadata(BaseModule):
     def split_data_by_batch(self, j_data):
         """Split metadata from json for each batch of samples found according to folder location of the samples.
         Args:
-            files_found_dict (dict): A dictionary containing file paths identified for each configuration item.
             j_data (list(dict)): List of dictionaries, one per sample, including metadata for that sample
         Returns:
             data_by_batch (dict(list(dict))): Dictionary containing parts of j_data corresponding to each
@@ -856,16 +808,16 @@ class BioinfoMetadata(BaseModule):
             ]
         return data_by_batch
 
-    def split_tables_by_batch(self, files_found_dict, sufix, batch_data, output_dir):
+    def split_tables_by_batch(self, files_found_dict, file_tag, batch_data, output_dir):
         """Filter table content to output a new table containing only the samples present in given metadata
         Args:
             files_found_dict (dict): A dictionary containing file paths identified for each configuration item.
-            sufix (str): Sufix to be added to the new table file name.
+            file_tag (str): File tag to be added to the new table file name.
             batch_data (list(dict)): Metadata corresponding to a single folder with samples (folder)
             output_dir (str): Output location for the generated tabular file
         """
 
-        def extract_batch_rows_to_file(file, sufix):
+        def extract_batch_rows_to_file(file, new_filename):
             """Create a new table file only with rows matching samples in batch_data"""
             extdict = {".csv": ",", ".tsv": "\t", ".tab": "\t"}
             file_extension = os.path.splitext(file)[1]
@@ -876,8 +828,6 @@ class BioinfoMetadata(BaseModule):
             file_df[sample_col] = file_df[sample_col].astype(str)
             file_df = file_df[file_df[sample_col].isin(batch_samples)]
 
-            base, ext = os.path.splitext(os.path.basename(file))
-            new_filename = f"{base}_{sufix}{ext}"
             os.makedirs(os.path.join(output_dir, "analysis_results"), exist_ok=True)
             output_path = os.path.join(output_dir, "analysis_results", new_filename)
             file_df.to_csv(output_path, index=False, sep=extdict.get(file_extension))
@@ -893,7 +843,15 @@ class BioinfoMetadata(BaseModule):
             sample_colpos = self.get_sample_idx_colpos(key)
             for file in files:
                 try:
-                    extract_batch_rows_to_file(file, sufix)
+                    if self.software_config[key].get("filepath_name"):
+                        filepath_key = self.software_config[key].get("filepath_name")
+                        new_filename = os.path.basename(
+                            list({row[filepath_key] for row in batch_data})[0]
+                        )
+                    else:
+                        base, ext = os.path.splitext(os.path.basename(file))
+                        new_filename = f"{base}_{file_tag}{ext}"
+                    extract_batch_rows_to_file(file, new_filename)
                 except Exception as e:
                     if self.software_config[key].get("required"):
                         log_type = "error"
@@ -957,79 +915,16 @@ class BioinfoMetadata(BaseModule):
                             prev_metadata_dict[sample_id] = item
                             stderr.print(f"[green]Sample '{sample_id}' updated (auto).")
                     else:
-                        stderr.print(
-                            f"[red]Sample '{sample_id}' has different data in {batch_filepath} and new metadata. Can't merge."
-                        )
-                        self.log.error(
-                            "Sample %s has different data in %s and new metadata. Can't merge.",
-                            sample_id,
-                            batch_filepath,
-                        )
-                        sys.exit(1)
+                        errtxt = f"Sample '{sample_id}' has different data in {batch_filepath} and new metadata. Can't merge."
+                        stderr.print(f"[red]{errtxt}")
+                        self.log.error(errtxt)
+                        raise ValueError(errtxt)
             else:
                 prev_metadata_dict[sample_id] = item
 
         merged_metadata = list(prev_metadata_dict.values())
         relecov_tools.utils.write_json_to_file(merged_metadata, batch_filepath)
         return merged_metadata
-
-    def save_merged_files(self, files_dict, file_tag, output_dir=None):
-        """
-        Process and save files that where split by cod and that have a function to be processed
-
-        Args:
-            files_dict (dict): A dictionary containing file paths identified for each configuration item.
-            file_tag (str): Tag that will be used for output filenames includes batch date (same as download date) and hex.
-            output_dir (str): Path to save output files generated during processing.
-
-        Returns:
-            None
-        """
-        method_name = f"{self.save_merged_files.__name__}"
-        for key, config in self.software_config.items():
-            func_name = config.get("function")
-            # Skip configurations that do not match the conditions
-            if config.get("split_by_batch") is True and func_name:
-                try:
-                    file_path = files_dict[key]
-                    stderr.print(f"[blue]Processing splitted file: {file_path}")
-                    self.log.info(f"Processing splitted file: {file_path}")
-                except KeyError:
-                    self.update_all_logs(
-                        method_name,
-                        "warning",
-                        f"No file path found for '{file_path}'",
-                    )
-                    continue
-                try:
-                    # Dynamically import the function from the specified module
-                    if func_name.startswith("utils/"):
-                        utils_name = "relecov_tools.assets.pipeline_utils.utils"
-                        func_name = func_name.split("/", 1)[1]
-                    else:
-                        utils_name = (
-                            f"relecov_tools.assets.pipeline_utils.{self.software_name}"
-                        )
-
-                    import_statement = f"import {utils_name}"
-                    exec(import_statement)
-                    # Get method name and execute it.
-                    data = eval(
-                        utils_name
-                        + "."
-                        + func_name
-                        + "(file_path, file_tag, '"
-                        + self.software_name
-                        + "', output_dir)"
-                    )
-                except Exception as e:
-                    self.update_all_logs(
-                        self.save_merged_files.__name__,
-                        "error",
-                        f"Error occurred while parsing '{func_name}': {e}.",
-                    )
-                    sys.exit(self.log_report.print_log_report(method_name, ["error"]))
-        return data
 
     def get_multiple_sample_files(self):
         multiple_sample_files = []
@@ -1060,6 +955,20 @@ class BioinfoMetadata(BaseModule):
                     sample.pop(k)
         return data
 
+    def split_extra_json_data(self, extra_data, batch_data):
+        sample_names_in_batch = {
+            sample_data["sequencing_sample_id"] for sample_data in batch_data
+        }
+        filtered_batch_data = [
+            sample_data
+            for sample_data in extra_data
+            if sample_data["sample_name"] in sample_names_in_batch
+        ]
+
+        filename = filtered_batch_data[0]["file_name"]
+
+        return filtered_batch_data, filename
+
     def create_bioinfo_file(self):
         """Create the bioinfodata json with collecting information from lab
         metadata json, mapping_stats, and more information from the files
@@ -1069,17 +978,131 @@ class BioinfoMetadata(BaseModule):
             bool: True if the bioinfo file creation process was successful.
         """
 
+        tag = "bioinfo_lab_metadata_"
+        file_tag = self.batch_id + "_" + self.hex
+        out_filename = tag + file_tag + ".json"
+        batch_filename = self.tag_filename(out_filename)
+
+        year = str(datetime.now().year)
+        out_path = os.path.join(self.output_dir, year)
+        os.makedirs(out_path, exist_ok=True)
+        batch_filepath = os.path.join(out_path, out_filename)
+
         # Check samplesheet for matching samples
         self.validate_samplenames()
 
         # Find and validate bioinfo files
         stderr.print("[blue]Scanning input directory...")
+        self.log.info("Scanning input directory")
         files_found_dict = self.scann_directory()
         stderr.print("[blue]Validating required files...")
+        self.log.info("Validating required files")
         self.validate_software_mandatory_files(files_found_dict)
+        stderr.print("[blue]Adding bioinfo metadata to read lab metadata...")
+        self.log.info("Adding bioinfo metadata to read lab metadat")
+        self.j_data, extra_json_data = self.add_bioinfo_results_metadata(
+            files_found_dict, file_tag, self.j_data, out_path
+        )
+        stderr.print("[blue]Adding fixed values")
+        self.log.info("Adding fixed values")
+        self.j_data = self.add_fixed_values(self.j_data, batch_filename)
+        # Adding files path
+        stderr.print("[blue]Adding files path to read lab metadata")
+        self.log.info("Adding files path to read lab metadata")
+        self.j_data = self.add_bioinfo_files_path(files_found_dict, self.j_data)
+
+        # Dynamically import the function from the specified module
+        module = importlib.import_module(
+            f"relecov_tools.assets.pipeline_utils.{self.software_name}"
+        )
+        try:
+            if hasattr(module, "quality_control_evaluation"):
+                qc_func = getattr(module, "quality_control_evaluation")
+                self.j_data = qc_func(self.j_data)
+
+        except (AttributeError, NameError, TypeError, ValueError) as e:
+            self.update_all_logs(
+                self.create_bioinfo_file.__name__,
+                "warning",
+                f"Could not evaluate quality_control_evaluation for batch {self.j_data}: {e}",
+            )
+            stderr.print(
+                f"[orange]Could not evaluate quality_control_evaluation for batch {self.j_data}: {e}"
+            )
+
+        # Filter properties from batch_data that are not included in the schema
+        self.j_data = self.filter_properties(self.j_data)
+        valid_rows, invalid_rows = relecov_tools.validate.Validate.validate_instances(
+            self.j_data, self.json_schema, "sequencing_sample_id"
+        )
+        valid_samples = [sample.get("sequencing_sample_id") for sample in valid_rows]
+        for sample in valid_samples:
+            self.logsum.feed_key(key=out_path, sample=sample)
+        if len(invalid_rows) > 0:
+            unique_failed_samples = list(
+                set(
+                    sample
+                    for samples in invalid_rows["samples"].values()
+                    for sample in samples
+                )
+            )
+            for error_message, failed_samples in invalid_rows["samples"].items():
+                num_samples = len(failed_samples)
+                field_with_error = invalid_rows["fields"][error_message]
+                sample_list = "', '".join(failed_samples)
+                error_text = f"{error_message} in field '{field_with_error}' for {num_samples} sample/s: '{sample_list}'"
+                if len(unique_failed_samples) == len(self.j_data):
+                    self.logsum.add_error(key=out_path, entry=error_text)
+                else:
+                    self.logsum.add_warning(key=out_path, entry=error_text)
+                self.log.info(error_text)
+                stderr.print(f"[red]{error_text}")
+
+                for failsamp in failed_samples:
+                    self.logsum.add_error(
+                        key=out_path, sample=failsamp, entry=error_text
+                    )
+
+            if not self.soft_validation:
+                self.parent_create_error_summary(
+                    called_module="read-bioinfo-metadata", logs=self.logsum.logs
+                )
+                errtxt = "Metadata was not completely validate, fix the errors or run with --soft_validation"
+                self.log.warning(errtxt)
+                stderr.print(f"[red]{errtxt}")
+                return False
+
+        else:
+            stderr.print("[green]Bioinfo json succesfully validated.")
+            self.log.info("Bioinfo json succesfully validated.")
+
+        self.j_data = valid_rows
+
+        for sample in self.j_data:
+            self.logsum.feed_key(
+                key=out_path, sample=sample.get("sequencing_sample_id")
+            )
+
         # Split files found based on each batch of samples
         data_by_batch = self.split_data_by_batch(self.j_data)
-        sufix = datetime.now().strftime("%Y%m%d%H%M%S")
+
+        if os.path.exists(batch_filepath):
+            stderr.print(
+                f"[blue]Bioinfo metadata {batch_filepath} file already exists. Merging new data if possible."
+            )
+            self.log.info(
+                "Bioinfo metadata %s file already exists. Merging new data if possible."
+                % batch_filepath
+            )
+            self.j_data = self.merge_metadata(batch_filepath, self.j_data)
+        else:
+            relecov_tools.utils.write_json_to_file(self.j_data, batch_filepath)
+
+        self.log.info("Created output json file: %s" % batch_filepath)
+        stderr.print(f"[green]Created batch json file: {batch_filepath}")
+
+        self.log.info("Splitting data by batch")
+        stderr.print("[blue]Splitting data by batch")
 
         # Add bioinfo metadata to j_data
         for batch_dir, batch_dict in data_by_batch.items():
@@ -1099,48 +1122,18 @@ class BioinfoMetadata(BaseModule):
                 "submitting_institution_id", batch_dir.split("/")[-2]
             )
             batch_date = first_sample.get("batch_id", batch_dir.split("/")[-1])
-            self.set_batch_id(batch_date)
-            self.logsum.feed_key(batch_dir)
             file_tag = batch_date + "_" + self.hex
             stderr.print(f"[blue]Processing data from {batch_dir}")
-            stderr.print("[blue]Adding bioinfo metadata to read lab metadata...")
-            self.split_tables_by_batch(files_found_dict, sufix, batch_data, batch_dir)
-            batch_data = self.add_bioinfo_results_metadata(
-                files_found_dict, batch_data, sufix, file_tag, batch_dir
+            self.log.info(f"Processing data from {batch_dir}")
+
+            self.split_tables_by_batch(
+                files_found_dict, file_tag, batch_data, batch_dir
             )
-            stderr.print("[blue]Adding software versions to read lab metadata...")
-            stderr.print("[blue]Adding fixed values")
-            batch_data = self.add_fixed_values(batch_data)
-            # Adding files path
-            stderr.print("[blue]Adding files path to read lab metadata")
-            batch_data = self.add_bioinfo_files_path(files_found_dict, batch_data)
+
             tag = "bioinfo_lab_metadata_"
             batch_filename = tag + lab_code + ".json"
             batch_filename = self.tag_filename(batch_filename)
             batch_filepath = os.path.join(batch_dir, batch_filename)
-            module = eval(f"relecov_tools.assets.pipeline_utils.{self.software_name}")
-            try:
-                if hasattr(module, "quality_control_evaluation"):
-                    qc_func = eval(
-                        f"relecov_tools.assets.pipeline_utils.{self.software_name}.quality_control_evaluation"
-                    )
-                    qc_data = qc_func(batch_data)
-                    for sample in batch_data:
-                        sample_id = sample.get("sequencing_sample_id")
-                        if sample_id in qc_data:
-                            sample.update(qc_data[sample_id])
-            except (AttributeError, NameError, TypeError, ValueError) as e:
-                self.update_all_logs(
-                    self.create_bioinfo_file.__name__,
-                    "warning",
-                    f"Could not evaluate quality_control_evaluation for batch {batch_dir}: {e}",
-                )
-                stderr.print(
-                    f"[orange]Could not evaluate quality_control_evaluation for batch {batch_dir}: {e}"
-                )
-
-            # Filter properties from batch_data that are not included in the schema
-            batch_data = self.filter_properties(batch_data)
 
             if os.path.exists(batch_filepath):
                 stderr.print(
@@ -1160,45 +1153,18 @@ class BioinfoMetadata(BaseModule):
             self.log.info("Created output json file: %s" % batch_filepath)
             stderr.print(f"[green]Created batch json file: {batch_filepath}")
 
-        year = str(datetime.now().year)
-        out_path = os.path.join(self.output_dir, year)
-        os.makedirs(out_path, exist_ok=True)
+            for extra_json in extra_json_data:
+                filtered_batch_data, filename = self.split_extra_json_data(
+                    extra_json, batch_data
+                )
 
-        stderr.print("[blue]Saving previously splitted files to output directory")
-        batch_dates = []
-        # Get batch date for all the samples
-        for batch_dir, batch_dict in data_by_batch.items():
-            if batch_dir.split("/")[-1] not in batch_dates:
-                batch_dates.append(batch_dir.split("/")[-1])
+                extra_filename = filename + "_" + lab_code + "_" + file_tag + ".json"
+                extra_filepath = os.path.join(batch_dir, extra_filename)
 
-        if len(batch_dates) == 1:
-            batch_date = str(batch_dates[0])
-        else:
-            stderr.print(
-                "[orange]More than one batch date in the same json data. Using current date as batch date."
-            )
-            self.log.info(
-                "More than one batch date in the same json data. Using current date as batch date."
-            )
-            batch_date = self.basemod_date
-        self.set_batch_id(batch_date)
-        file_tag = self.batch_id + "_" + self.hex
-        self.save_merged_files(files_found_dict, file_tag, out_path)
-        batch_filename = self.tag_filename("bioinfo_lab_metadata" + ".json")
-        stderr.print("[blue]Writting output json file")
-        file_path = os.path.join(out_path, batch_filename)
-        if os.path.exists(file_path):
-            stderr.print(
-                f"[blue]Bioinfo metadata {file_path} file already exists. Merging new data if possible."
-            )
-            self.log.info(
-                "Bioinfo metadata %s file already exists. Merging new data if possible."
-                % file_path
-            )
-            batch_data = self.merge_metadata(file_path, self.j_data)
-        else:
-            relecov_tools.utils.write_json_to_file(self.j_data, file_path)
-        stderr.print(f"[green]Sucessful creation of bioinfo analyis file: {file_path}")
+                relecov_tools.utils.write_json_to_file(
+                    filtered_batch_data, extra_filepath
+                )
+
         self.parent_create_error_summary(
             called_module="read-bioinfo-metadata", logs=self.logsum.logs
         )
