@@ -17,7 +17,7 @@ stderr = rich.console.Console(
 )
 
 
-class SftpRelecov:
+class SftpClient:
     """Class to handle SFTP connection with remote server. It uses paramiko library to establish
     the connection. The class can be used to upload and download files from the remote server.
     The class can be initialized with a configuration file and with the username and password.
@@ -32,7 +32,7 @@ class SftpRelecov:
     """
 
     def __init__(self, conf_file=None, username=None, password=None):
-        if conf_file is None:
+        if not conf_file:
             config_json = ConfigJson()
             self.sftp_server = config_json.get_topic_data("sftp_handle", "sftp_server")
             self.sftp_port = config_json.get_topic_data("sftp_handle", "sftp_port")
@@ -202,6 +202,11 @@ class SftpRelecov:
                 return True
             except FileNotFoundError as e:
                 log.error("Unable to fetch file %s ", e)
+                try:
+                    os.remove(destination)
+                except OSError:
+                    log.error(f"Could not delete {destination} after failed fetch")
+                    pass
                 return False
 
     @reconnect_if_fail(n_times=3, sleep_time=30)
@@ -294,9 +299,39 @@ class SftpRelecov:
             self.sftp.put(local_path, remote_file)
             return True
         except FileNotFoundError as e:
-            log.error("File not found %s", e)
-            stderr.print("[red]File not found")
+            log.error(f"Could not upload file {local_path}: {e}")
+            stderr.print(f"[red]Could not upload file {local_path}: {e}")
             return False
+
+    @reconnect_if_fail(n_times=3, sleep_time=30)
+    def copy_within_sftp(self, src_path, dest_path, buffer_size=65536):
+        """
+        Copies a file within the SFTP server by reading and writing in blocks.
+
+        Args:
+            src_path (str): Path to the source file on the SFTP server.
+            dest_path (str): Path where the file should be copied to on the SFTP server.
+            buffer_size (int, optional): Block size for reading/writing in bytes. Default is 64 KB.
+
+        Returns:
+            bool: True if the copy was successful, False if it was not
+        """
+        try:
+            log.info(f"Copying file within SFTP: {src_path} -> {dest_path}")
+            with self.sftp.open(src_path, "rb") as src_file:
+                with self.sftp.open(dest_path, "wb") as dest_file:
+                    while True:
+                        data = src_file.read(buffer_size)
+                        if not data:
+                            break
+                        dest_file.write(data)
+            log.info(f"File successfully copied within SFTP to: {dest_path}")
+            return True
+        except FileNotFoundError:
+            log.error(f"Source file not found: {src_path}")
+        except IOError as e:
+            log.error(f"Error during SFTP copy operation: {e}")
+        return False
 
     @reconnect_if_fail(n_times=3, sleep_time=30)
     def close_connection(self):
